Clients, Missions, MissionLines Request & CRUD

+++ factures/database/migrations/2022_04_04_131730_create_missions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('reference');
             $table->integer('client_id');
-            $table->float('advance');
+            $table->integer('advance');
         });
     }
